third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-workers-view-expected.txt

Tests ServiceWorkersView on resources panel.

Select ServiceWorkers tree element.
Register ServiceWorker for scope1
==== ServiceWorkersView ====
http://127.0.0.1:8000/inspector/service-workers/resources/scope1/
URL
/inspector/service-workers/resources/service-worker-empty.js
State
running
Updated
RESPONSE-TIME
Last-Modified
LAST-MODIFIED
Recent messages
LAST-MODIFIED
RESPONSE-TIME

============================
Register ServiceWorker for scope2
==== ServiceWorkersView ====
http://127.0.0.1:8000/inspector/service-workers/resources/scope2/
URL
/inspector/service-workers/resources/service-worker-empty.js
State
running
Updated
RESPONSE-TIME
Last-Modified
LAST-MODIFIED
Recent messages
LAST-MODIFIED
RESPONSE-TIME
URL
/inspector/service-workers/resources/service-worker-empty.js
State
running
Updated
RESPONSE-TIME
Last-Modified
LAST-MODIFIED
Recent messages
LAST-MODIFIED
RESPONSE-TIME

============================
Unegister ServiceWorker for scope1
==== ServiceWorkersView ====
http://127.0.0.1:8000/inspector/service-workers/resources/scope1/ - deleted
URL
/inspector/service-workers/resources/service-worker-empty.js
State
runninginspect
Updated
RESPONSE-TIME
Last-Modified
LAST-MODIFIED
Recent messages
LAST-MODIFIED
RESPONSE-TIME
URL
/inspector/service-workers/resources/service-worker-empty.js
State
running
Updated
RESPONSE-TIME
Last-Modified
LAST-MODIFIED
Recent messages
LAST-MODIFIED
RESPONSE-TIME

============================
Unegister ServiceWorker for scope1
